requirements.txt
google-generativeai
Pillow
python-dotenv
requests
pytest